+++ client/src/upload.cc
@@ -7,7 +7,7 @@ int main(int argc, char *argv[]) {
   int fd = open(argv[1], O_RDONLY);
   struct stat file_state;
   fstat(fd, &file_state);
-  size_t file_size = file_state.st_size;
+  uint64_t file_size = file_state.st_size;
   if (file_size > SMALL_FILE_SIZE_MAX) {
     do_big_file_upload(fd, argv[1], file_size);
   } else {
